virtual_sim_firebase.py
# 📱 Virtual SIM Layer with Firebase Integration

## 🔧 Building a Digital SIM Solution

You're right that we can create a virtual SIM layer that functions like a real SIM. Here's how to build it with Firebase integration:

## 🛠️ Architecture Overview

```
[Mobile App] ←→ [Firebase] ←→ [Virtual SIM Layer] ←→ [Web Interface]
     ↑                              ↓
     └──────────────────────[SDR Hardware/NIC] 
```

## 🔥 Firebase Integration Implementation

### 1. Firebase Configuration
```javascript
// firebase-config.js
import { initializeApp } from 'firebase/app';
import { getAuth, signInWithEmailAndPassword } from 'firebase/auth';
import { getFirestore, collection, addDoc, onSnapshot } from 'firebase/firestore';

const firebaseConfig = {
  apiKey: "your-api-key",
  authDomain: "your-project.firebaseapp.com",
  projectId: "your-project-id",
  storageBucket: "your-project.appspot.com",
  messagingSenderId: "123456789",
  appId: "1:123456789:web:abcdef123456"
};

const app = initializeApp(firebaseConfig);
const auth = getAuth(app);
const db = getFirestore(app);

export { auth, db };
```

### 2. Virtual SIM Service
```python
# virtual_sim_service.py
import firebase_admin
from firebase_admin import credentials, firestore
import jwt
import hashlib
from datetime import datetime

class VirtualSIMService:
    def __init__(self):
        # Initialize Firebase Admin SDK
        if not firebase_admin._apps:
            cred = credentials.Certificate('firebase-service-account.json')
            firebase_admin.initialize_app(cred)
        
        self.db = firestore.client()
        self.sim_cards = {}  # In-memory SIM registry
    
    def register_virtual_sim(self, user_id, phone_number, encryption_key):
        """Register a new virtual SIM"""
        sim_id = hashlib.sha256(f"{user_id}{phone_number}".encode()).hexdigest()[:16]
        
        sim_data = {
            'sim_id': sim_id,
            'user_id': user_id,
            'phone_number': phone_number,
            'encryption_key': encryption_key,
            'registered_at': datetime.utcnow(),
            'status': 'active',
            'imsi': f"00101{hashlib.md5(sim_id.encode()).hexdigest()[:10]}",
            'ki': hashlib.sha256(encryption_key.encode()).hexdigest()
        }
        
        # Store in Firebase
        self.db.collection('virtual_sims').document(sim_id).set(sim_data)
        self.sim_cards[sim_id] = sim_data
        
        return sim_id
    
    def authenticate_sim(self, sim_id, challenge):
        """Authenticate virtual SIM (SIM authentication protocol)"""
        if sim_id not in self.sim_cards:
            return None
            
        sim_data = self.sim_cards[sim_id]
        # Simplified A3/A8 algorithm simulation
        response = hashlib.sha256(f"{challenge}{sim_data['ki']}".encode()).hexdigest()[:16]
        kc = hashlib.sha256(f"{response}{sim_data['ki']}".encode()).hexdigest()[:16]
        
        return {
            'response': response,
            'kc': kc,
            'imsi': sim_data['imsi']
        }
    
    def send_sms_via_firebase(self, from_sim, to_number, message):
        """Send SMS through Firebase messaging"""
        sms_data = {
            'from': from_sim,
            'to': to_number,
            'message': message,
            'timestamp': datetime.utcnow(),
            'status': 'sent',
            'via': 'virtual_sim'
        }
        
        # Store in Firebase
        doc_ref = self.db.collection('sms_messages').add(sms_data)
        
        # Also store in user's message history
        user_doc = self.db.collection('user_messages').document(from_sim)
        user_doc.set({
            'messages': firestore.ArrayUnion([{
                'id': doc_ref[1].id,
                'data': sms_data
            }])
        }, merge=True)
        
        return doc_ref[1].id
    
    def listen_for_incoming_sms(self, sim_id, callback):
        """Listen for incoming SMS via Firebase"""
        def on_snapshot(docs, changes, read_time):
            for change in changes:
                if change.type.name == 'ADDED':
                    data = change.document.to_dict()
                    if data.get('to') == self.sim_cards.get(sim_id, {}).get('phone_number'):
                        callback(data)
        
        # Listen to SMS collection
        self.db.collection('sms_messages').where('to', '==', 
            self.sim_cards.get(sim_id, {}).get('phone_number')).on_snapshot(on_snapshot)

# Usage example
virtual_sim_service = VirtualSIMService()