requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
motor==3.3.2
pymongo==4.6.1
pyjwt==2.8.0
pydantic==2.5.3
python-multipart==0.0.6
websockets==12.0
